// frontend/src/generated-api-client/model/get-all-quizzes-sort-order.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Quiziset API
 * The RESTful APIs Docs for the Quiziset app
 * OpenAPI spec version: 1.0
 */

export type GetAllQuizzesSortOrder = (typeof GetAllQuizzesSortOrder)[keyof typeof GetAllQuizzesSortOrder]

export const GetAllQuizzesSortOrder = {
  asc: 'asc',
  desc: 'desc',
} as const